HACK: intel/tools/dump_gpu: Enable device local memory on ATS.
On ATS, we need to enable the bit on this register to make sure the GGTT
entry bit for local/system memory actually works, otherwise it will be
ignored.

Since we are setting the GGTT entries to point to device local
(discrete) memory already, and then using
AUB_MEM_TRACE_MEMORY_ADDRESS_SPACE_GGTT to write into the memory mapped
by those, we need to enable local memory too.