!UPSTREAM: fixup i965/miptree: Fix aux surface bo offset.
On i965, usually we create one BO for the surface, and one for the aux
surface. But if we support the CCS modifier for the render buffer, we
create a single BO for both.

Thus, using mt->aux_buf->bo->gtt_offset points to the same address as
the main surface, since they use the same bo. We need to add the aux_buf
offset too, which will be 0 everywhere else except that render buffer,
and things should still work.