!UPSTREAM: i965/miptree: Prepare the miptree before accessing it.
Most intel_miptree_map_* functions will call intel_miptree_access_raw()
or intel_miptree_prepare_access() if a resolve is needed. However,
intel_miptree_map_map() and intel_miptree_map_tiled_memcpy() don't do
that, since they are used by the other functions.

If we want to fallback to them in the case where we don't support
_map_blit() or _map_movntdqa(), we need to first resolve the surface,
otherwise any mapping/reading that we do from it will be wrong. This
fallback will happen when we disable LLC, which is not supported on ATS.

This commit wraps them with a _map_cpu() function that does the right
thing.