iris: Emit 3DSTATE_BINDING_TABLE_POOL_ALLOC for gen12+.
On gen12+, Binding Table Pointers are an offset relative to the Surface
State Base Address anymore. Instead, they are relative to the State
Binding Table Pool Address, which is set by the command above.

So let's make sure we emit that command (pointing to the same address as
the Surface State Base Addresss), and everything should stay working as
before.

Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>
[jordan.l.justen@intel.com: iris port]
Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>