anv: rework device queue storage
Previously we things a bit hard coded because we only even dealt with
one engine.

With ATS we start seeing variable number of compute, so store the
number of queue for each family on the physical device.

Also store the queue family & instances in each queue so that we know
exactly what engine is addressed.

v2: Deal with kernel not supporting listing engine instances (Jordan)

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>